merges master in android with VC++ fixes
Former-commit-id: 3a32d8426b83a068e3feb109eb3a34953b0dfbc9

+++ src/smp/smp.cpp
@@ -31,7 +31,7 @@ void smp::Sleep(core::Time seconds)
 	auto microseconds = core::SecondsToDuration<std::chrono::microseconds>(seconds);
 	std::this_thread::sleep_for(microseconds);
 #else
-	Uint32 ms = 1000 * seconds;
+	Uint32 ms = Uint32(1000. * seconds);
 	SDL_Delay(ms);
 #endif
 }
